Release #19044#
* fix(ci)

* fix(ci): release simplified

* fix(ci): an infinite loop fixed in tests with mariadbd restart

* fix(delivery): add delivery to old yum repo (#456)

* enh(ci): rebuild yum metadata only one time (#463)

* fix(ci): fix workflow trigger on push dev-22.10.x (#466)

* enh(packaging): add rpm sign to workflow 2210 (#468)

* enh(ci): add repository_name and debian delivery 2210 (#476) (#478)

* enh(ci): add repository_name and delivery (#476)

* enh(ci): add branch stability and delivery

* remove duplicate code

* enable delivery tests

* add deb delivery debug

* add deb delivery debug

* fix credentials

* disable deb delivery tests & debug

* fix(ci): update debian delivery url for 22.10

* fix(broker/lua): repeated fields should be ok with streamconnector

* chore(doc): changelog updated
* cleanup(broker/lua): refacto of the code
* cleanup(broker/lua): deadcode removed

REFS: MON-15903

* enh(ci): remove buster from build and delivery (#471) (#472)

* enh(ci): remove buster from build and delivery

* remove buster dockerfile

* feat(chore): add Deepsource configuration file (#479)

* use RELEASE instead of DISTRIB in package naming (#481) (#484)

* fix(cmake): conan path added to cmake (#505)

* fix(ci): registry updated

* enh(ci): Add rpm delivery cleanup and structure (#499) 2210 (#500)

* enh(ci): Add rpm delivery cleanup and structure (#499)

* enh(ci): Add rpm delivery cleanup and structure

* Add version to delivery inputs

* Add condition to use sshkey for rpm delivery only

* Update delivery action

* Update action delivery

* Update action delivery

* Update action delivery

* Update action delivery

* Update action delivery

* Enable cleanup

* Update action.yml

* Update action.yml

* Update action.yml

* fix(ci): correct docker registry link (#494)

* fix(ci): correct docker registry link

* Update action.yml

Fix run folder naming

* fix(conan): path updated to find protoc


* Handle deb stable delivery conditionally 2210 (#508)

* fix(ci): handle debian stable delivery conditionally 2210

* Update debian delivery url

* enh(ci): remove paths to workflow files in workflows for components (#513) (#523)

* chore(license): update maintainer and email addresses (#532)

* Mon 16454 add alma 8 to centreon collect nightly tests (#506) (#519)

* enh(ci): Add alma8 to centreon-collect nightly

* Add job to robot-nightly workflow

* enable ci test branch

* Update robot nightly workflow

* add debug to get-packages

* Add release to github_env

* add delivery test

* Get pkg dir version name

* Fix url

* fix url

* make curl silent

* fix curl

* add mariadb client

* update get-packages

* disable nightly on alma 8

* update alma8 image

* Add MariaDB packages to alma8 dockerfile

* remove extra sudo

* update collect-test-robot.sh

* Remove duplicate Robot test job

* fix(ci/tests): attempt to upgrade grpc python library

* Install proper grpc version based on OS version id

* help cmake to find protoc

* disable delivery tests

* remove unused lines from actions

* Remove more unused lines


* chore(ci): Update delivery credentials (#543) (#544)

* chore(ci): Update delivery credentials (#543)

* Update delivery credentials

* Update delivery credentials and inputs

* Update deb delivery credentials

* Update .github/workflows/centreon-collect.yml


---------


* chore(ci): Add extra delivery to artifactory (#553)

* MON-17043 cache-name-rpmbuild shorten (#573)

REFS:MON-17043

* chore(ci): move to harbor docker registry (22.10) (#581)

* chore(license): update package licenses (#559)

* enh(ci): handle debian delivery failure 2210 (#603)

* enh(cmake): new cmake.sh version

* chore(ci): fix cache name

* chore(ci): fix cache name

* Update robot-nightly.yml

Disable nightly workflow

---------

+++ ci/scripts/collect-test-robot.sh
@@ -1,21 +1,21 @@
 #!/bin/bash
 set -e
+set -x
 
 export RUN_ENV=docker
 
-echo "########################### configure and start sshd ############################"
-ssh-keygen -t rsa -f /etc/ssh/ssh_host_rsa_key
-ssh-keygen -t ecdsa -f /etc/ssh/ssh_host_ecdsa_key
-ssh-keygen -t ed25519 -f /etc/ssh/ssh_host_ed25519_key
+echo "########################### Configure and start sshd ###########################"
+ssh-keygen -t rsa -f /etc/ssh/ssh_host_rsa_key -P ""
+ssh-keygen -t ecdsa -f /etc/ssh/ssh_host_ecdsa_key -P ""
+ssh-keygen -t ed25519 -f /etc/ssh/ssh_host_ed25519_key -P ""
 /usr/sbin/sshd > /dev/null 2>&1 &
 
-echo "########################### start mariadb ############################"
+echo "########################### Start MariaDB ######################################"
 mariadbd --user=root > /dev/null 2>&1 &
 sleep 5
 
-echo "########################### init centreon database ############################"
+echo "########################### Init centreon database ############################"
 
-cd /src/tests/
 mysql -e "CREATE USER IF NOT EXISTS 'centreon'@'localhost' IDENTIFIED BY 'centreon';"
 
 mysql -e "GRANT SELECT,UPDATE,DELETE,INSERT,CREATE,DROP,INDEX,ALTER,LOCK TABLES,CREATE TEMPORARY TABLES, EVENT,CREATE VIEW ON *.* TO  'centreon'@'localhost';"
@@ -23,41 +23,43 @@ mysql -e "GRANT SELECT,UPDATE,DELETE,INSERT,CREATE,DROP,INDEX,ALTER,LOCK TABLES,
 mysql -u centreon -pcentreon < resources/centreon_storage.sql
 mysql -u centreon -pcentreon < resources/centreon.sql
 
-echo "########################### build and install centreon collect ############################"
+echo "########################## Install centreon collect ###########################"
 
-rm -rf /src/build
-mkdir /src/build
-cd /src/build/
-DISTRIB=$(lsb_release -rs | cut -f1 -d.)
-if [ "$DISTRIB" = "7" ] ; then
-    source /opt/rh/devtoolset-9/enable
-fi 
-conan install .. -s compiler.cppstd=14 -s compiler.libcxx=libstdc++11 --build=missing
-if [ $(cat /etc/issue | awk '{print $1}') = "Debian" ] ; then
-    CXXFLAGS="-Wall -Wextra" cmake -DCMAKE_INSTALL_PREFIX=/usr -DCMAKE_BUILD_TYPE=Debug -DWITH_USER_BROKER=centreon-broker -DWITH_USER_ENGINE=centreon-engine -DWITH_GROUP_BROKER=centreon-broker -DWITH_GROUP_ENGINE=centreon-engine -DWITH_TESTING=On -DWITH_MODULE_SIMU=On -DWITH_BENCH=On -DWITH_CREATE_FILES=OFF ..
-else 
-    CXXFLAGS="-Wall -Wextra" cmake3 -DCMAKE_INSTALL_PREFIX=/usr -DCMAKE_BUILD_TYPE=Debug -DWITH_USER_BROKER=centreon-broker -DWITH_USER_ENGINE=centreon-engine -DWITH_GROUP_BROKER=centreon-broker -DWITH_GROUP_ENGINE=centreon-engine -DWITH_TESTING=On -DWITH_MODULE_SIMU=On -DWITH_BENCH=On -DWITH_CREATE_FILES=OFF ..
-fi
+echo "Installation..."
+/usr/bin/rpm -Uvvh --force --nodeps *.rpm
 
-#Build
-make -j9
-make -j9 install
-
-echo "########################### install robot framework ############################"
+echo "########################### Install Robot Framework ###########################"
 cd /src/tests/
 pip3 install -U robotframework robotframework-databaselibrary pymysql python-dateutil
 
-yum install "Development Tools" python3-devel -y
+yum groupinstall "Development Tools" -y
+yum install python3-devel -y
 
-pip3 install grpcio==1.33.2 grpcio_tools==1.33.2
+# Get OS version id
+VERSION_ID=$(grep '^VERSION_ID' /etc/os-release | sed -En 's/^VERSION_ID="([[:digit:]])\.[[:digit:]]"/\1/p')
+
+# Force version for el7 only
+if [ -f /etc/os-release ]; then
+    case "$VERSION_ID" in
+        7)
+            pip3 install grpcio==1.33.2 grpcio_tools==1.33.2
+            ;;
+        8)
+            pip3 install grpcio grpcio_tools
+            ;;
+        *)
+            echo "OS Version is neither 7 or 8"
+            ;;
+    esac
+fi
 
 ./init-proto.sh
 
-echo "########################### run centreon collect test robot ############################"
+echo "####################### Run Centreon Collect Robot Tests #######################"
 cd /src/tests/
 robot --nostatusrc .
 
-echo "########################### generate folder report ############################"
+echo "########################### Generate Folder Report #############################"
 mkdir reports
 cp log.html output.xml report.html reports
